+++ sdk/c4-client.ts
@@ -2,8 +2,8 @@
 // Generated from /openapi.json via src/openapi-sdk-gen.js.
 // Do not edit by hand — re-run `c4 openapi --sdk` to refresh.
 
-// Spec version: 1.10.39
-// Generated at: 2026-05-01T17:46:21.406Z
+// Spec version: 1.10.40
+// Generated at: 2026-05-01T17:50:43.245Z
 
 export interface postAuthLoginBody {
   user: string; /** Username */